Solution_Accelerators/Advanced_RAG/src/evals/rag_eval/requirements.txt
pandas==2.2.3
argparse==1.4.0
azure-ai-ml==1.25.0
azure-ai-evaluation==1.2.0
azure-identity==1.20.0
azureml-core==1.59.0.post1
azureml-fsspec==1.3.1
tqdm==4.67.1
transformers==4.48.3
torch==2.6.0
promptflow==1.17.2
promptflow-azure==1.17.2
promptflow-tools==1.5.0
pylint==3.3.4
ruff==0.9.6
pytest==8.3.4
keyring==24.2.0
artifacts-keyring==0.4.0
twine==6.1.0
build==1.2.2.post1
marshmallow==3.23.2
azure-keyvault-secrets==4.7.0
opencensus-ext-azure==1.1.9
opencensus==0.11.2